+++ src/aleatora/wav.py
@@ -13,13 +13,13 @@ def load_array(filename, resample=False):
         sample_rate = w.getframerate()
         buffer = w.readframes(w.getnframes())
     if width == 2:
-        data = np.frombuffer(buffer, dtype=np.int16).reshape(-1, channels).astype(np.float) / np.iinfo(np.int16).max
+        data = np.frombuffer(buffer, dtype=np.int16).reshape(-1, channels).astype(np.float32) / np.iinfo(np.int16).max
     elif width == 3:
         audio = np.frombuffer(buffer, dtype=np.uint8).reshape(-1, 3)
         converted = np.bitwise_or.reduce(audio << np.array([8, 16, 24]), dtype=np.int32, axis=1)
-        data = converted.astype(np.float).reshape(-1, channels) / np.iinfo(np.int32).max
+        data = converted.astype(np.float32).reshape(-1, channels) / np.iinfo(np.int32).max
     elif width == 4:
-        data = np.frombuffer(buffer, dtype=np.int32).astype(np.float).reshape(-1, channels) / np.iinfo(np.int32).max
+        data = np.frombuffer(buffer, dtype=np.int32).astype(np.float32).reshape(-1, channels) / np.iinfo(np.int32).max
     else:
         raise NotImplementedError(f"{width*8}-bit wave files not supported")
     if resample and streams.SAMPLE_RATE != w.getframerate():
@@ -46,7 +46,7 @@ def save(comp, filename, chunk_size=16384, verbose=False):
     w.setnchannels(channels)
     w.setsampwidth(2)
     w.setframerate(streams.SAMPLE_RATE)
-    chunk = np.empty((chunk_size, channels), dtype=np.float)
+    chunk = np.empty((chunk_size, channels), dtype=np.float32)
     siter = iter(comp)
     # Avoid holding onto memory if e.g. memoize() is involved:
     del comp
